feat: add nickname and rename desc to common
